penambahan pengecekan bank soal pada jadwal ujian yg belum terdaftar / maba belum terdaftar dengan jadwal tersebut sebelum di hapus

+++ UnpakCbt.BankSoal.Application/BankSoal/DeleteBankSoal/DeleteBankSoalCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
+﻿
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,11 +10,13 @@ using UnpakCbt.Common.Domain;
 using UnpakCbt.Modules.BankSoal.Application.Abstractions.Data;
 using UnpakCbt.Modules.BankSoal.Application.BankSoal.CreateBankSoal;
 using UnpakCbt.Modules.BankSoal.Domain.BankSoal;
+using UnpakCbt.Modules.JadwalUjian.PublicApi;
 
 namespace UnpakCbt.Modules.BankSoal.Application.BankSoal.DeleteBankSoal
 {
     internal sealed class DeleteBankSoalCommandHandler(
     IBankSoalRepository bankSoalRepository,
+    IJadwalUjianApi JadwalUjianApi,
     IUnitOfWork unitOfWork,
     ILogger<DeleteBankSoalCommand> logger)
     : ICommandHandler<DeleteBankSoalCommand>
@@ -26,6 +29,12 @@ namespace UnpakCbt.Modules.BankSoal.Application.BankSoal.DeleteBankSoal
             {
                 logger.LogError("BankSoal dengan referensi uuid {uuid} tidak ditemukan", request.uuid);
                 return Result.Failure(BankSoalErrors.NotFound(request.uuid));
+            }
+
+            int registeredExamScheduleCount = await JadwalUjianApi.CountByBankSoalHasRegisteredMabaAsync(request.uuid, cancellationToken);
+            if (registeredExamScheduleCount>0) {
+                logger.LogError("BankSoal dengan referensi uuid {uuid} sudah terdaftar di jadwal ujian aktif", request.uuid);
+                return Result.Failure(BankSoalErrors.CommandAbort(request.uuid, "Delete"));
             }
 
             await bankSoalRepository.DeleteAsync(existingBankSoal!);
